+++ eslint.config.mjs
@@ -4,26 +4,36 @@ import simpleImportSortPlugin from 'eslint-plugin-simple-import-sort';
 import tseslint from 'typescript-eslint';
 
 export default tseslint.config(
-  {
-    files: ['**/*.ts'],
-    plugins:{
-      ['simple-import-sort']: simpleImportSortPlugin,
-    }
-  },
   { ignores: ['sampleWorkspace', 'out', 'dist', '**/*.d.ts'] },
   eslint.configs.recommended,
   ...tseslint.configs.recommended,
   eslintPluginPrettierRecommended,
   {
+    files: ['**/*.ts'],
+    plugins: {
+      'simple-import-sort': simpleImportSortPlugin,
+    },
     rules: {
       '@typescript-eslint/no-unused-vars': [
         'warn',
-        { varsIgnorePattern: '^_', argsIgnorePattern: '^_' },
+        {
+          varsIgnorePattern: '^_',
+          argsIgnorePattern: '^_',
+          caughtErrorsIgnorePattern: '^_',
+          destructuredArrayIgnorePattern: '^_',
+        },
       ],
       '@typescript-eslint/no-explicit-any': 'warn',
       'simple-import-sort/imports': 'warn',
       'simple-import-sort/exports': 'warn',
-      'prettier/prettier': 'warn'
+      'prettier/prettier': 'warn',
+    },
+  },
+  // 测试目录特殊规则：允许使用 any 类型
+  {
+    files: ['src/test/**/*.ts', '__mocks__/**/*.ts'],
+    rules: {
+      '@typescript-eslint/no-explicit-any': 'off',
     },
   },
 );