+++ vite.config.ts
@@ -18,8 +18,15 @@ export default defineConfig({
     ],
   },
   test: {
+    globals: true,
+    environment: 'node',
     include: ['src/test/unit/**/*.spec.ts'],
+    typecheck: {
+      tsconfig: './tsconfig.test.json',
+    },
     coverage: {
+      provider: 'v8',
+      reporter: ['text', 'json', 'html', 'lcov'],
       exclude: ['node_modules', 'out', 'src/test', 'src/typings', '.vscode-test'],
     },
   },